+++ components/collector/tests/source_collectors/azure_devops/base.py
@@ -42,7 +42,11 @@ class AzureDevopsJobsTestCase(AzureDevopsTestCase):
                 "path": self.path,
                 "name": "include_pipeline",
                 "_links": {"web": {"href": f"{self.url}/build"}},
-                "latestCompletedBuild": {"result": "failed", "finishTime": "2019-10-15T12:24:10.1905868Z"},
+                "latestCompletedBuild": {
+                    "result": "failed",
+                    "startTime": "2019-10-15T12:20:10.1905868Z",
+                    "finishTime": "2019-10-15T12:24:10.1905868Z",
+                },
             },
             {
                 "path": self.path,
@@ -74,6 +78,7 @@ class AzureDevopsJobsTestCase(AzureDevopsTestCase):
                 "key": self.pipeline.replace("/", "-"),
                 "url": f"{self.url}/build",
                 "build_date": "2019-10-15 12:24:10.190586+00:00",
+                "build_duration": "4",
                 "build_status": "failed",
             },
         ]
@@ -130,6 +135,7 @@ class AzureDevopsPipelinesTestCase(AzureDevopsTestCase):
                 "key": f"{self.test_pipeline['id']}-20191015_1",  # safe_entity_key
                 "url": f"{self.url}/_build/results?buildId=1",
                 "build_date": "2019-10-15 12:24:10.190586+00:00",
+                "build_duration": "4",
                 "build_result": "succeeded",
                 "build_status": "completed",
             },
@@ -139,6 +145,7 @@ class AzureDevopsPipelinesTestCase(AzureDevopsTestCase):
                 "key": f"{self.test_pipeline['id']}-20191015_2",  # safe_entity_key
                 "url": f"{self.url}/_build/results?buildId=2",
                 "build_date": "2019-10-15 12:34:10.190586+00:00",
+                "build_duration": "4",
                 "build_result": "succeeded",
                 "build_status": "completed",
             },