+++ components/collector/tests/source_collectors/azure_devops/test_source_up_to_dateness.py
@@ -22,7 +22,11 @@ class AzureDevopsSourceUpToDatenessTest(AzureDevopsTestCase):
                     "path": r"\\folder",
                     "name": "pipeline",
                     "_links": {"web": {"href": f"{self.url}/build"}},
-                    "latestCompletedBuild": {"result": "failed", "finishTime": self.timestamp},
+                    "latestCompletedBuild": {
+                        "result": "failed",
+                        "startTime": self.timestamp,
+                        "finishTime": self.timestamp,
+                    },
                 },
             ],
         }